Refactoring) store.jsx와 reducer.jsx의 분리
